refactor(daemon): extract the endpoint probe from the health checker
The rule that decides who may serve on the endpoint needs to classify what is
serving it, but that classifier lived in daemon-health, which already depends on
the ownership module — importing it back would cycle. No behaviour change.

Also records that a non-socket occupying the endpoint reports ENOTSOCK on macOS
and ECONNREFUSED on Linux, so both must be read as "nothing can serve this".
